# docs/swagger.yaml
basePath: /
definitions:
  auth.ErrorResponse:
    properties:
      message:
        example: Invalid param ID
        type: string
    type: object
...
  model.LoginFiberResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  model.ResultMessage:
    properties:
      en:
        type: string
      ru:
        type: string
      tk:
        type: string
    type: object
  model.Success:
    properties:
      message:
        type: string
    type: object
...
  model.UserApplication:
    properties:
      activity_field_id:
        type: integer
      address:
        type: string
      company_name:
        type: string
      company_type_id:
        type: integer
      email:
        type: string
      full_name:
        type: string
      licence_expiry_date:
        type: string
      licence_issue_date:
        type: string
      phone:
        type: string
      role_id:
        description: 1 user, 2 dealer, 3 logistic, 4 broker, 5 car service
        type: integer
      vat_number:
        type: string
    required:
    - activity_field_id
    - address
    - company_name
    - company_type_id
    - email
    - full_name
    - licence_expiry_date
    - licence_issue_date
    - phone
    - role_id
    - vat_number
    type: object
  model.UserEmailConfirmationRequest:
    properties:
      email:
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
  model.UserLoginEmail:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  model.UserLoginGoogle:
    properties:
      token_id:
        type: string
    required:
    - token_id
    type: object
  model.UserLoginPhone:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
  model.UserPhoneConfirmationRequest:
    properties:
      otp:
        type: string
      phone:
        type: string
    required:
    - otp
    - phone
    type: object
host: 84.200.87.48
info:
  contact: {}
  description: Project Description
  title: Project name
  version: "1.0"
paths:
...
  /api/v1/auth/account/{id}:
    delete:
      description: Deletes the authenticated user's account and related data
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - auth
  /api/v1/auth/send-application:
    post:
      consumes:
      - application/json
      description: Sends an application to the database
      parameters:
      - description: Application
        in: body
        name: application
        required: true
        schema:
          $ref: '#/definitions/model.UserApplication'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginFiberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      summary: Send application
      tags:
      - auth
  /api/v1/auth/send-application-document:
    post:
      consumes:
      - application/json
      description: Sends application documents to the database
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      - description: A PDF document file
        in: formData
        name: licence
        required: true
        type: file
      - description: A PDF document file
        in: formData
        name: memorandum
        required: true
        type: file
      - description: A PDF document file
        in: formData
        name: copy_of_id
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      summary: Application documents
      tags:
      - auth
  /api/v1/auth/user-email-confirmation:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: User email confirmation credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserEmailConfirmationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginFiberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      summary: User email confirmation
      tags:
      - auth
  /api/v1/auth/user-login-email:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: User login email credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserLoginEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      summary: User login email
      tags:
      - auth
  /api/v1/auth/user-login-google:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: User login google credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserLoginGoogle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginFiberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      summary: User login google
      tags:
      - auth
  /api/v1/auth/user-login-phone:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: User login phone credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserLoginPhone'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      summary: User login
      tags:
      - auth
  /api/v1/auth/user-phone-confirmation:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: User phone confirmation credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserPhoneConfirmationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginFiberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResultMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResultMessage'
      summary: User phone confirmation
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"